+++ changelog.md
@@ -1,6 +1,6 @@
 # Version history of weather-information-collector
 
-## Version 0.?.? (2022-10-??)
+## Version 0.10.1 (2022-01-01)
 
 An informational notice will be shown when there is at least one task uses the
 DarkSky API to get data. DarkSky will shut down on 31st March 2023, so users